`@0x/contracts-staking`: Change naming conventions of internal variables. `@0x/contracts-staking`: Make `StakingPoolRewardVault.depositFor()` take explicit operator and members reward amounts. `@0x/contracts-staking`: Remove storing of `operatorShare` for pools in `StakingPoolRewardVault`. `@0x/contracts-staking`: Remove `deposit()` function in `StakingPoolRewardVault` (just use fallback fn). `@0x/contracts-staking`: Make `operatorShare` denominated in `1 / 10**18`. `@0x/contracts-staking`: Remove `MixinPanic` (maybe another PR). `@0x/contracts-staking`: Calculate operator/members split in `MixinExchangeFees` before depositing into reward vault. `@0x/contracts-staking`: Make quantity fields in `IStakingPoolRewardVault.Balance` struct uint256 (previously uint96).

+++ contracts/staking/contracts/src/interfaces/IStructs.sol
@@ -51,10 +51,11 @@ interface IStructs {
 
     /// @dev State for Staking Pools (see MixinStakingPool).
     /// @param operatorAddress Address of pool operator.
-    /// @param operatorShare Portion of pool rewards owned by operator.
+    /// @param operatorShare Portion of pool rewards owned by operator, denominated
+    ///                      as 1 / `TOKEN_MULTIPLIER`.
     struct Pool {
         address payable operatorAddress;
-        uint8 operatorShare;
+        uint256 operatorShare;
     }
 
     /// @dev State for a pool that actively traded during the current epoch.
